fix(stella-tui): keep deck_render.rs under its file-size ratchet
Moves the CACHE/SAVED/WARMTH statline integration tests (and their local
fixtures) from deck_render.rs's test module into cache_panel.rs, where they
belong thematically and where there's headroom under the 1500-line new-file
cap. deck_render.rs was over its allowlisted 2005+50 limit at 2064 lines;
now at 1905. render_status_bar is pub(crate) so cache_panel's tests can
render a full statline frame.

Also widens the full-frame ethos-chip fixture from 160 to 190 cols — the
new SAVED/WARMTH cells no longer left room for the chip at the old width.
